+++ google/src/main/scala/org/broadinstitute/dsde/workbench/google/HttpGoogleIamDAO.scala
@@ -1,11 +1,14 @@
 package org.broadinstitute.dsde.workbench.google
 
+import java.time.Instant
+import java.time.format.DateTimeFormatter
+
 import akka.actor.ActorSystem
 import akka.http.scaladsl.model.StatusCodes
+import cats.data.OptionT
 import cats.instances.future._
 import cats.instances.list._
 import cats.instances.map._
-import cats.data.OptionT
 import cats.syntax.foldable._
 import cats.syntax.functor._
 import cats.syntax.semigroup._
@@ -17,15 +20,16 @@ import com.google.api.client.http.HttpResponseException
 import com.google.api.client.json.jackson2.JacksonFactory
 import com.google.api.services.cloudresourcemanager.CloudResourceManager
 import com.google.api.services.cloudresourcemanager.model.{Binding => ProjectBinding, Policy => ProjectPolicy, SetIamPolicyRequest => ProjectSetIamPolicyRequest}
-import com.google.api.services.iam.v1.model.{CreateServiceAccountRequest, ServiceAccount, Binding => ServiceAccountBinding, Policy => ServiceAccountPolicy, SetIamPolicyRequest => ServiceAccountSetIamPolicyRequest}
+import com.google.api.services.iam.v1.model.{CreateServiceAccountKeyRequest, CreateServiceAccountRequest, ServiceAccount, Binding => ServiceAccountBinding, Policy => ServiceAccountPolicy, SetIamPolicyRequest => ServiceAccountSetIamPolicyRequest}
 import com.google.api.services.iam.v1.{Iam, IamScopes}
-import org.broadinstitute.dsde.workbench.metrics.GoogleInstrumentedService
-import org.broadinstitute.dsde.workbench.model._
 import org.broadinstitute.dsde.workbench.google.HttpGoogleIamDAO._
 import org.broadinstitute.dsde.workbench.google.model._
+import org.broadinstitute.dsde.workbench.metrics.GoogleInstrumentedService
+import org.broadinstitute.dsde.workbench.model._
 
 import scala.collection.JavaConverters._
 import scala.concurrent.{ExecutionContext, Future}
+import scala.util.Try
 
 /**
   * Created by rtitle on 10/2/17.
@@ -147,6 +151,38 @@ class HttpGoogleIamDAO(serviceAccountClientId: String,
         executeGoogleRequest(request)
       }.void
     }
+  }
+
+  override def createServiceAccountKey(serviceAccountProject: GoogleProject, serviceAccountEmail: WorkbenchUserServiceAccountEmail): Future[WorkbenchUserServiceAccountKey] = {
+    val request = new CreateServiceAccountKeyRequest()
+      .setPrivateKeyType("TYPE_GOOGLE_CREDENTIALS_FILE")
+      .setKeyAlgorithm("KEY_ALG_RSA_2048")
+    val creater = iam.projects().serviceAccounts().keys().create(s"projects/${serviceAccountProject.value}/serviceAccounts/${serviceAccountEmail.value}", request)
+    retryWhen500orGoogleError { () =>
+      executeGoogleRequest(creater)
+    } map { key =>
+      WorkbenchUserServiceAccountKey(
+        WorkbenchUserServiceAccountKeyId(key.getName.split('/').last),
+        WorkbenchUserServiceAccountPrivateKeyData(key.getPrivateKeyData),
+        Option(key.getValidAfterTime).flatMap(googleTimestampToInstant),
+        Option(key.getValidBeforeTime).flatMap(googleTimestampToInstant))
+    }
+  }
+
+  override def removeServiceAccountKey(serviceAccountProject: GoogleProject, serviceAccountEmail: WorkbenchUserServiceAccountEmail, keyId: WorkbenchUserServiceAccountKeyId): Future[Unit] = {
+    val request = iam.projects().serviceAccounts().keys().delete(s"projects/${serviceAccountProject.value}/serviceAccounts/${serviceAccountEmail.value}/keys/${keyId.value}")
+    retryWithRecoverWhen500orGoogleError{ () =>
+      executeGoogleRequest(request)
+      ()
+    } {
+      case e: HttpResponseException if e.getStatusCode == StatusCodes.NotFound.intValue => ()
+    }
+  }
+
+  private def googleTimestampToInstant(googleTimestamp: String): Option[Instant] = {
+    Try {
+      Instant.from(DateTimeFormatter.ISO_INSTANT.parse(googleTimestamp))
+    }.toOption
   }
 
   private def getProjectPolicy(googleProject: GoogleProject): Future[Policy] = {