+++ model/src/main/scala/org/broadinstitute/dsde/workbench/model/WorkbenchIdentity.scala
@@ -1,5 +1,8 @@
 package org.broadinstitute.dsde.workbench.model
 
+import java.time.Instant
+import java.time.format.DateTimeFormatter
+
 import spray.json.{DefaultJsonProtocol, JsString, JsValue, RootJsonFormat}
 
 /**
@@ -19,6 +22,17 @@ object WorkbenchIdentityJsonSupport {
     def read(value: JsValue) = ???
   }
 
+  implicit object InstantFormat extends RootJsonFormat[Instant] {
+    def write(instant: Instant): JsString = {
+      JsString(DateTimeFormatter.ISO_INSTANT.format(instant))
+    }
+
+    def read(value: JsValue): Instant = value match {
+      case JsString(str) => Instant.from(DateTimeFormatter.ISO_INSTANT.parse(str))
+      case _ => throw new WorkbenchException(s"Unable to unmarshal Instant from $value")
+    }
+  }
+
   implicit val WorkbenchUserIdFormat = ValueObjectFormat(WorkbenchUserId)
   implicit val WorkbenchUserEmailFormat = ValueObjectFormat(WorkbenchUserEmail)
   implicit val WorkbenchUserFormat = jsonFormat2(WorkbenchUser)
@@ -31,6 +45,10 @@ object WorkbenchIdentityJsonSupport {
   implicit val WorkbenchUserPetServiceAccountEmailFormat = ValueObjectFormat(WorkbenchUserServiceAccountEmail)
   implicit val workbenchUserPetServiceAccountDisplayNameFormat = ValueObjectFormat(WorkbenchUserServiceAccountDisplayName)
   implicit val WorkbenchUserPetServiceAccountFormat = jsonFormat3(WorkbenchUserServiceAccount)
+
+  implicit val WorkbenchUserServiceAccountKeyIdFormat = ValueObjectFormat(WorkbenchUserServiceAccountKeyId)
+  implicit val WorkbenchUserServiceAccountPrivateKeyDataFormat = ValueObjectFormat(WorkbenchUserServiceAccountPrivateKeyData)
+  implicit val WorkbenchUserServiceAccountKeyFormat = jsonFormat4(WorkbenchUserServiceAccountKey)
 }
 
 sealed trait WorkbenchSubject
@@ -52,3 +70,7 @@ case class WorkbenchUserServiceAccountEmail(value: String) extends WorkbenchEmai
   def toAccountName: WorkbenchUserServiceAccountName = WorkbenchUserServiceAccountName(value.split("@")(0))
 }
 case class WorkbenchUserServiceAccountDisplayName(value: String) extends ValueObject //A friendly name.
+
+case class WorkbenchUserServiceAccountKeyId(value: String) extends ValueObject
+case class WorkbenchUserServiceAccountPrivateKeyData(value: String) extends ValueObject with Base64Support
+case class WorkbenchUserServiceAccountKey(id: WorkbenchUserServiceAccountKeyId, privateKeyData: WorkbenchUserServiceAccountPrivateKeyData, validAfter: Option[Instant], validBefore: Option[Instant])